Add: customer hook squeleton hook
